Revert "Removed another duplicate FHiCL file already in icaruscode."
It looks like it is not in icaruscode any more, either.
Despite being an art configuration, it is used also by non-art unit tests
which extract the current geometry configuration out of it.
The configuration of the equivalent art-based unit test and the
non-art unit tests are intentionally merged into this one file.

This reverts commit 0a35628ed6807940164bd37143957fc21bb738ab.